pytorch/requirements.txt
torch==2.6.0
matplotlib==3.10.1
numpy==2.2.3
pyyaml==6.0.2
tqdm==4.67.1